Static instance of SC until a fresh is needed
