+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_a_message_retry_audit_is_sent_to_a_remote_instance.cs
@@ -16,8 +16,8 @@
         [Test]
         public async Task Should_mark_as_resolved_on_master()
         {
-            SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
-            CustomInstanceConfiguration = ConfigureWaitingForMasterToSubscribe;
+            SetSetInstanceSettings(ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues);
+            SetCustomInstanceConfiguration(ConfigureWaitingForMasterToSubscribe);
 
             FailedMessage failure;
