+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_multiple_messages_are_emitted_by_a_saga.cs
@@ -40,7 +40,7 @@
         [Test]
         public async Task Saga_history_can_be_fetched_on_master()
         {
-            SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
+            SetSetInstanceSettings(ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues);
 
             SagaHistory sagaHistory = null;
