+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_processed_message_multi_instance_endpoint_by_messagetype.cs
@@ -21,7 +21,7 @@
         [Test]
         public async Task Should_be_found()
         {
-            SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
+            SetSetInstanceSettings(ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues);
 
             var response = new List<MessagesView>();
