+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_proessed_message_multi_instance_endpoint_by_messages.cs
@@ -21,7 +21,7 @@
         [Test]
         public async Task Should_be_found()
         {
-            SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
+            SetSetInstanceSettings(ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues);
 
             var response = new List<MessagesView>();
